Fix chart displaying
When dealing with results directly from Apollo Client they will be returned defined with object.preventExtensions
Vega doesn't like this and throws an error.

Instead it is better to copy the data into a simple const and then pass that through to the render component.
In general this seems nicer as you don't couple the render component to the data types defined for the containers.